refactor(providers): extract openai api protocol
