+++ self/subcortex/providers/src/protocols/openai-api/index.ts
@@ -1,0 +1,8 @@
+export {
+  chatCompletionsAdapter,
+  createChatCompletionsAdapter,
+} from './adapter.js';
+export {
+  CHAT_COMPLETIONS_PROVIDER_DEFINITION,
+  ChatCompletionsProvider,
+} from './provider.js';
