+++ self/subcortex/providers/src/providers/openai/definition.ts
@@ -1,3 +1,3 @@
 export {
   CHAT_COMPLETIONS_PROVIDER_DEFINITION as providerDefinition,
-} from '../../chat-completions-provider.js';
+} from '../../protocols/openai-api/provider.js';
