+++ self/subcortex/providers/src/providers/openai/index.ts
@@ -1,4 +1,4 @@
 export { providerAdapter } from './adapter.js';
 export { providerDefinition } from './definition.js';
 export { providerFactory } from './provider.js';
-export { ChatCompletionsProvider } from '../../chat-completions-provider.js';
+export { ChatCompletionsProvider } from '../../protocols/openai-api/provider.js';
