+++ self/subcortex/providers/src/shared/chat-completions-adapter.ts
@@ -1,194 +1,4 @@
-import type { TraceId } from '@nous/shared';
-import type { ParsedModelOutput } from './output.js';
-import {
-  defineProviderAdapter,
-  type AdapterCapabilities,
-  type AdapterFormatInput,
-  type AdapterFormattedRequest,
-  type ProviderAdapter,
-} from './adapter-types.js';
-
-const CHAT_COMPLETIONS_ADAPTER_CAPABILITIES: AdapterCapabilities = {
-  nativeToolUse: true,
-  cacheControl: false,
-  extendedThinking: false,
-  streaming: false,
-};
-
-export function createChatCompletionsAdapter(): ProviderAdapter {
-  return {
-    capabilities: CHAT_COMPLETIONS_ADAPTER_CAPABILITIES,
-    formatRequest(input: AdapterFormatInput): AdapterFormattedRequest {
-      const systemPrompt = Array.isArray(input.systemPrompt)
-        ? input.systemPrompt.join('\n\n')
-        : input.systemPrompt;
-
-      const messages = [
-        { role: 'system' as const, content: systemPrompt } as Record<string, unknown>,
-        ...input.context.map((frame) => {
-          // Assistant frame with tool_calls metadata → OpenAI tool_calls on assistant message
-          if (frame.role === 'assistant' && Array.isArray(frame.metadata?.tool_calls)) {
-            const toolCalls = (frame.metadata!.tool_calls as Array<{ id?: string; name: string; input: unknown }>)
-              .map((tc, index) => ({
-                id: tc.id ?? `call_${index}`,
-                type: 'function' as const,
-                function: {
-                  name: tc.name,
-                  arguments: (() => { try { return JSON.stringify(tc.input ?? {}); } catch { return '{}'; } })(),
-                },
-              }));
-            return {
-              role: 'assistant' as const,
-              content: frame.content,
-              tool_calls: toolCalls,
-            };
-          }
-          // Tool result with tool_call_id metadata → OpenAI tool result message
-          // Include `name` when present so the model can recognize which tool
-          // returned this result.
-          if (frame.role === 'tool' && frame.metadata?.tool_call_id) {
-            return {
-              role: 'tool' as const,
-              content: frame.content,
-              tool_call_id: frame.metadata.tool_call_id as string,
-              ...(frame.name ? { name: frame.name } : {}),
-            };
-          }
-          return {
-            role: frame.role === 'tool' ? ('user' as const) : frame.role,
-            content: frame.content,
-          };
-        }),
-      ];
-
-      const result: Record<string, unknown> = { messages };
-
-      // Map tool definitions to OpenAI tools format
-      if (input.toolDefinitions && input.toolDefinitions.length > 0) {
-        result.tools = input.toolDefinitions.map((tool) => ({
-          type: 'function',
-          function: {
-            name: tool.name,
-            description: tool.description ?? '',
-            parameters: tool.inputSchema ?? {},
-          },
-        }));
-      }
-
-      // ModelRequirements currently has profile + fallbackPolicy only. Pass the
-      // profile through as metadata until richer model parameters are available.
-      if (input.modelRequirements) {
-        result.model_profile = input.modelRequirements.profile;
-      }
-
-      return { input: result };
-    },
-    parseResponse(output: unknown, _traceId: TraceId): ParsedModelOutput {
-      try {
-        return parseChatCompletionsResponse(output);
-      } catch {
-        // Fallback to text-mode — never throw
-        return {
-          response: String(output ?? ''),
-          toolCalls: [],
-          memoryCandidates: [],
-          contentType: 'text',
-        };
-      }
-    },
-  };
-}
-
-function parseChatCompletionsResponse(output: unknown): ParsedModelOutput {
-  if (typeof output !== 'object' || output === null) {
-    return {
-      response: String(output ?? ''),
-      toolCalls: [],
-      memoryCandidates: [],
-      contentType: 'text',
-    };
-  }
-
-  const obj = output as Record<string, unknown>;
-
-  // Handle OpenAI chat completion response shape
-  // { choices: [{ message: { content, tool_calls } }] }
-  const choices = obj.choices;
-  if (Array.isArray(choices) && choices.length > 0) {
-    const message = (choices[0] as Record<string, unknown>)?.message;
-    if (message && typeof message === 'object') {
-      const msg = message as Record<string, unknown>;
-      const content = typeof msg.content === 'string' ? msg.content : '';
-      const toolCalls = parseChatCompletionsToolCalls(msg.tool_calls);
-      return {
-        response: content,
-        toolCalls,
-        memoryCandidates: [],
-        contentType: 'text',
-      };
-    }
-  }
-
-  // Handle direct message shape
-  if ('content' in obj || 'tool_calls' in obj) {
-    const content = typeof obj.content === 'string' ? obj.content : '';
-    const toolCalls = parseChatCompletionsToolCalls(obj.tool_calls);
-    return {
-      response: content,
-      toolCalls,
-      memoryCandidates: [],
-      contentType: 'text',
-    };
-  }
-
-  // Handle plain response field (our canonical format)
-  if ('response' in obj && typeof obj.response === 'string') {
-    return {
-      response: obj.response,
-      toolCalls: [],
-      memoryCandidates: [],
-      contentType: 'text',
-    };
-  }
-
-  return {
-    response: String(output),
-    toolCalls: [],
-    memoryCandidates: [],
-    contentType: 'text',
-  };
-}
-
-function parseChatCompletionsToolCalls(
-  toolCalls: unknown,
-): Array<{ name: string; params: unknown; id?: string }> {
-  if (!Array.isArray(toolCalls)) return [];
-  const result: Array<{ name: string; params: unknown; id?: string }> = [];
-  for (const tc of toolCalls) {
-    if (tc && typeof tc === 'object' && 'function' in tc) {
-      const tcObj = tc as Record<string, unknown>;
-      const fn = tcObj.function;
-      if (fn && typeof fn === 'object') {
-        const fnObj = fn as Record<string, unknown>;
-        const name = typeof fnObj.name === 'string' ? fnObj.name : '';
-        let params: unknown = {};
-        if (typeof fnObj.arguments === 'string') {
-          try { params = JSON.parse(fnObj.arguments); } catch { params = {}; }
-        }
-        const id = typeof tcObj.id === 'string' ? tcObj.id : undefined;
-        if (name) result.push({ name, params, id });
-      }
-    }
-  }
-  return result;
-}
-
-export const chatCompletionsAdapter = defineProviderAdapter({
-  adapterKey: 'chat-completions',
-  displayName: 'Chat Completions',
-  protocol: 'chat-completions',
-  capabilities: CHAT_COMPLETIONS_ADAPTER_CAPABILITIES,
-  create() {
-    return createChatCompletionsAdapter();
-  },
-});
+export {
+  chatCompletionsAdapter,
+  createChatCompletionsAdapter,
+} from '../protocols/openai-api/adapter.js';
